p2p-energy-blockchain-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
pydantic==2.5.1
requests==2.32.5
# Add solana or anchor client libraries later
# solana==0.28.0
# anchorpy==0.18.0
